Add files via upload
The n-queens puzzle is the problem of placing n queens on an n×n chessboard such that no two queens attack each other. Each solution contains a distinct board configuration of the n-queens' placement, where 'Q' and '.' both indicate a queen and an empty space respectively.
Genetic Algorithm is used with a novel fitness function as the Meta-heuristic. The aim of N-Queens Problem is to place N queens on an N x N chessboard, in a way so that no queen is in conflict with the others. Chromosome representation and genetic operations like Mutation and Crossover are described in detail. Results show that this approach yields promising and satisfactory results
in less time compared to that obtained from the previous approaches for several large values of N.